Make `Driver` a sealed trait.
The main reason for making it sealed is that it provides a limited
subset of behavior, making it a trait was mostly done for
the purposes of future expansion. So it is possible to replace the trait
implementation outright, if we want to add or remove options, or
otherwise change or provide alternate behaviors.

But i'm a bit skeptical that having these behaviors being specified
externally would be beneficial. It seems better that if you need a
specific behavior, make contact and we can discuss adding one.